Quaternions.jl

This package provides a bare-bones representation of attitude transformations through quaternions and transformation matrices (a.k.a. direction cosine matrices – DCM).

Installation

Just type:

using Pkg
Pkg.add(url="https://github.com/FraCpl/Quaternions.jl")

Notation and conventions

Given a 3-dimensional vector $v$, a reference frame $A$ and a reference frame $B$, we indicate with $v^A$ the projection of $v$ into $A$, and with $v^B$ its projection in $B$, so that the two projections can be related through the transformation matrix $R_{AB}$ as follows

\[v^A = R_{AB} v^B\]

The same transformation can be represented using the quaternion $q_{AB}$

\[v^A = q_{AB} ⊗ v^B ⊗ q_{AB}^*\]

where in the formula above, 3D vectors are represented by quaternions having a null scalar component.

In this package, quaternions are represented as 4-dimensional vectors, using the basic vector format natively provided by Julia. The scalar component of the quaternion corresponds to its first element, i.e., q[1]. We use a right-handed passive convention for representing attitude transformations through quaternions, and use $\theta_{AB}$ to indicate the angle by which the frame $A$ needs to be actively rotated to make it coincide with frame $B$ (i.e., the rotation from $A$ to $B$).

Note

Other (and weirder) quaternions notations exist in the litterature, which are obviously not covered by this package.

API

Quaternions

Quaternions.q_build — Method

q_build(qs,qv)

Build a quaternion from its scalar and vectorial components.

Quaternions.q_derivative — Method

q̇_AB = q_derivative(q_AB,ωAB_B)

Compute the time derivative of a unitary quaternion, given the corresponding angular velocity vector.

Mathematically, this function performs the following operation:

\[q̇_{AB} = \frac{1}{2} q_{AB} ⊗ [0; ω^B_{AB}]\]

where ωAB_B represents the angular velocity of frame $B$ with respect to frame $A$, projected into frame $B$.

Quaternions.q_fromAxes — Method

q_AB = q_fromAxes(xB_A, yB_A, zB_A)

Compute the attitude quaternion given as input the axes of a reference frame.

Quaternions.q_fromAxisAngle — Method

q_AB = q_fromAxisAngle(u,θ_AB)

Compute the unitary quaternion given as input an axis-angle representation.

Quaternions.q_fromDcm — Method

q_AB = q_fromDcm(R_AB)

Translate the input rotation matrix into a unitary quaternion.

Quaternions.q_identity — Method

q_identity()

Get identity quaternion, with scalar component equal to 1 and vector components equal to zero.

Quaternions.q_inverse — Method

q⁻¹ = q_inverse(q)

Compute the inverse of the input quaternion.

Quaternions.q_multiply — Method

q_AC = q_multiply(q_AB,q_BC)

Mutliply the two input quaternions as follows:

\[q_{AC} = q_{AB} ⊗ q_{BC}\]

Quaternions.q_multiplyn — Method

q = q_multiplyn(q1,q2,q3,...)

\[ q = q₁ ⊗ q₂ ⊗ q₃ ⊗ ...\]

Quaternions.q_random — Method

q_random()

Generate a random unitary quaternion.

Quaternions.q_toDcm — Method

R_AB = q_toDcm(q_AB)

Translate the input unitary quaternion into a transformation matrix.

\[R_{AB}(q_{AB}) = I + 2qₛ[qᵥ×] + 2[qᵥ×]²\]

Quaternions.q_transformVector — Method

v_A = q_transformVector(q_AB,v_B)

(previously known as q_rotateVector)

Project the vector v from frame B into frame A using the following passive rotation formula

\[vᴬ = q_{AB} ⊗ vᴮ ⊗ q_{BA}\]

In the formula above, 3D vectors are represented by quaternions having a null scalar component.

Quaternions.q_transpose — Method

q' = q_transpose(q)

Transpose the input quaternion.

Transformation matrices

Quaternions.dcm_fromAxisAngle — Method

R_AB = dcm_fromAxisAngle(u,θ_AB)

Compute the transformation matrix given the axis and angle.

\[R_{AB}(θ_{AB}) = I + \sin(θ_{AB})[u×] + (1 - \cos(θ_{AB}))[u×]^2\]

Quaternions.dcm_fromQuaternion — Method

R_AB = dcm_fromQuaternion(q_AB)

Compute a transformation matrix from a quaternion.

Quaternions.dcm_random — Method

dcm_random()

Generate a random transformation matrix.

Quaternions.dcm_toQuaternion — Method

q_AB = dcm_toQuaternion(R_AB)

Translate a transformation matrix into a quaternion.

Auxiliary functions

Quaternions.crossMat — Method

[v×] = crossMat(v)

Compute the cross product matrix of a vector v.

Index

Quaternions.crossMat

Quaternions.dcm_fromAxisAngle

Quaternions.dcm_fromQuaternion

Quaternions.dcm_random

Quaternions.dcm_toQuaternion

Quaternions.q_build

Quaternions.q_derivative

Quaternions.q_fromAxes

Quaternions.q_fromAxisAngle

Quaternions.q_fromDcm

Quaternions.q_identity

Quaternions.q_inverse

Quaternions.q_multiply

Quaternions.q_multiplyn

Quaternions.q_random

Quaternions.q_toDcm

Quaternions.q_transformVector

Quaternions.q_transpose